requirements.txt
numpy==2.3.0
openai==1.86.0
Pillow==11.2.1
pydantic==2.11.7
python-dotenv==1.1.0
Requests==2.32.4
scikit_learn==1.7.0
streamlit==1.45.1
